fixed a ton of logic errors with the filter searching
the older method worked for simple searches, but failed for more complex ones (specifically failing when you tried to use multiple AND filters on the same category. ie. trying to find books that were both action and adventure resulted in no books being shown).

this new method works for every test case i can come up with, but is extremely sloppy. i expect to come back to it and clean it up in the future